Client window size fix
AdjustWindowRect (and AdjustWindowRectEx) already include the necessary system metrics like borders, title bar height (SM_CYCAPTION), and so on  based on the window style you pass in. So manually adding them is not needed.
We create the rectangle specify the resolution from the .ini file then we ask AdjustWindowRect to essentially "simulate" the window frame size ahead of time, before actually creating the window — and that’s exactly the intended use case for AdjustWindowRect we then pass the same parameters to the CreateWindow function.